+++ processor/src/main/java/dz/jtsgen/processor/TsGenProcessor.java
@@ -22,9 +22,13 @@ package dz.jtsgen.processor;
 
 import dz.jtsgen.annotations.TSModule;
 import dz.jtsgen.annotations.TypeScript;
+import dz.jtsgen.processor.helper.ElementHelper;
 import dz.jtsgen.processor.jtp.TSModuleHandler;
 import dz.jtsgen.processor.jtp.TSModuleInfoEnforcer;
-import dz.jtsgen.processor.jtp.conv.*;
+import dz.jtsgen.processor.jtp.conv.JavaTypeProcessor;
+import dz.jtsgen.processor.jtp.conv.TSProcessingInfo;
+import dz.jtsgen.processor.jtp.conv.TSProcessingInfoBuilder;
+import dz.jtsgen.processor.jtp.conv.TypeScriptAnnotationProcessor;
 import dz.jtsgen.processor.model.TypeScriptModel;
 import dz.jtsgen.processor.renderer.TSRenderer;
 
@@ -36,10 +40,13 @@ import javax.lang.model.SourceVersion;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.TypeElement;
 import javax.tools.Diagnostic;
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 import java.util.logging.Logger;
-import java.util.stream.Collectors;
 
+import static dz.jtsgen.processor.helper.ElementHelper.typeOf;
 import static dz.jtsgen.processor.jtp.helper.RoundEnvHelper.filteredTypeSriptElements;
 import static dz.jtsgen.processor.model.TypeScriptModel.newModelWithDefaultModule;
 import static java.util.logging.Level.INFO;
@@ -106,19 +113,16 @@ public class TsGenProcessor extends AbstractProcessorWithLogging {
      * post process in case, only TSModule is defined and types to convert are defined in TSModule annotation
      */
     private void postProcess(Set<? extends TypeElement> annotations) {
-        if (annotations.size() == 1 && annotations.stream().anyMatch(this::isTsModuleAnnotation)) {
+        if (annotations.size() == 1 && annotations.stream().anyMatch(x -> ElementHelper.typeOf(x, TSModule.class))) {
             LOG.fine("P: Start PostPressing");
 //            this.typeScriptModel.getModuleInfo().additionalTypes().stream().map( )
             // ignore classes with TSIgnore
-            final TSProcessingInfo preliminaryProcessingInfo = TSProcessingInfoBuilder.of(this.processingEnv, this.typeScriptModel) ;
-            Set<Element> additionalElements = elementsFromStringList(this.typeScriptModel.getModuleInfo().additionalTypes(), preliminaryProcessingInfo);
-
+            final TSProcessingInfo processingInfo = TSProcessingInfoBuilder.of(this.processingEnv, this.typeScriptModel);
+            final JavaTypeProcessor handler = new TypeScriptAnnotationProcessor(processingInfo);
             // this is needed for updating data from CLI and calculating a name space mapping, if needed
-            new TSModuleInfoEnforcer(this.processingEnv, this.typeScriptModel).createUpdatedTSModuleInfo(additionalElements).ifPresent( x -> {
+            new TSModuleInfoEnforcer(this.processingEnv, this.typeScriptModel).createUpdatedTSModuleInfo(processingInfo.additionalTypesToConvert()).ifPresent(x -> {
                 typeScriptModel.addModuleInfo(x);
-
-                final JavaTypeProcessor handler = new TypeScriptAnnotationProcessor(preliminaryProcessingInfo);
-                handler.processElements(additionalElements);
+                handler.processElements(processingInfo.additionalTypesToConvert());
             });
 
         } else {
@@ -126,32 +130,21 @@ public class TsGenProcessor extends AbstractProcessorWithLogging {
         }
     }
 
-    private Set<Element> elementsFromStringList(List<String> elements, final TSProcessingInfo preliminaryProcessingInfo) {
-        return elements.stream()
-                .map( x -> preliminaryProcessingInfo.elementCache().typeElementByCanonicalName(x))
-                .filter(Objects::nonNull)
-                .collect(Collectors.toSet());
-    }
-
     // process all annotations
     private void processElements(TypeElement annotation, RoundEnvironment roundEnv) {
-        LOG.info( () -> String.format("P: Processing Annotation %s", annotation.getSimpleName()));
-        if (annotation.getSimpleName().contentEquals(TypeScript.class.getSimpleName())) {
+        LOG.info(() -> String.format("P: Processing Annotation %s", annotation.getSimpleName()));
+        if (typeOf(annotation, TypeScript.class)) {
             processTypeScriptAnnotation(annotation, roundEnv);
-        } else if (isTsModuleAnnotation(annotation)) {
+        } else if (typeOf(annotation, TSModule.class)) {
             Set<? extends Element> annotatedElements = roundEnv.getElementsAnnotatedWith(TSModule.class);
             if (annotatedElements.size() > 1) annotatedElements.forEach(
-                    x -> this.processingEnv.getMessager().printMessage(Diagnostic.Kind.WARNING,"Multiple TSModule not supported. Multiple Modules with same ", x));
+                    x -> this.processingEnv.getMessager().printMessage(Diagnostic.Kind.WARNING, "Multiple TSModule not supported. Multiple Modules with same ", x));
             new TSModuleHandler(this.processingEnv)
                     .process(annotatedElements)
                     .stream()
                     .findFirst()
                     .ifPresent(this.typeScriptModel::addModuleInfo);
         }
-    }
-
-    private boolean isTsModuleAnnotation(TypeElement annotation) {
-        return annotation.getSimpleName().contentEquals(TSModule.class.getSimpleName());
     }
 
     // process TypeScript Annotation this is after processing @TSModule
@@ -161,15 +154,14 @@ public class TsGenProcessor extends AbstractProcessorWithLogging {
         Set<Element> annotatedElements = filteredTypeSriptElements(roundEnv);
 
         // this is needed for updating data from CLI and calculating a name space mapping, if needed
-        new TSModuleInfoEnforcer(this.processingEnv,this.typeScriptModel).createUpdatedTSModuleInfo(annotatedElements).ifPresent( x -> {
+        new TSModuleInfoEnforcer(this.processingEnv, this.typeScriptModel).createUpdatedTSModuleInfo(annotatedElements).ifPresent(x -> {
             typeScriptModel.addModuleInfo(x);
 
-            final TSProcessingInfo TSProcessingInfo = TSProcessingInfoBuilder.of( this.processingEnv, typeScriptModel) ;
+            final TSProcessingInfo TSProcessingInfo = TSProcessingInfoBuilder.of(this.processingEnv, typeScriptModel);
             final JavaTypeProcessor handler = new TypeScriptAnnotationProcessor(TSProcessingInfo);
             handler.processAnnotations(roundEnv);
         });
     }
-
 
 
     public SourceVersion getSupportedSourceVersion() {