+++ processor/src/main/java/dz/jtsgen/processor/helper/Sets.java
@@ -18,14 +18,29 @@
  *
  */
 
+package dz.jtsgen.processor.helper;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Set;
 
 /**
- * Testing additionalTypes without TypeScript annotation
+ * Some helper functions for Sets
  */
-@TSModule(
-    moduleName = "additional_test",
-    additionalTypes = {"java.net.InetAddress"}
-)
-package jts.modules.additional;
+public final class Sets {
 
-import dz.jtsgen.annotations.TSModule;
+    /**
+     *
+     * @param first the first collection of the union set
+     * @param second the second collection of the union set
+     * @param <U> Type of the union elements
+     * @return an unmodifiable union set
+     */
+    public static <U> Set<U> union(Collection<U> first, Collection<U> second) {
+        final Set<U> result = new HashSet<U>();
+        result.addAll(first);
+        result.addAll(second);
+        return Collections.unmodifiableSet(result);
+    }
+}