+++ doc/source/Changelog.rst
@@ -9,6 +9,7 @@ New Features
 
 - Added documentation for Maven and Kotlin projects
 - Added Support for type bounds on classes and interfaces
+- Added support converting types without an annotation
 
 
 `jtsgen-0.3.0 <https://github.com/dzuvic/jtsgen/tree/jtsgen-0.3.0>`__ (2017-10-30)